# [Dynamic Spec Decoding] Minor fix for disabling speculative decoding