In the current implementation, even if `running_queue_size >= speculative_disable_by_batch_size`, it will still go through the speculative decoding logic, which includes get_spec_proposals with k=0, score_proposal, rejection sampling and create sampler output. The flow introduces extra overhead (especially rejection sampling), which makes disabling speculative decoding slower than 'real' without speculative decoding. 

To fix this, we can just reuse the `_run_no_spec` to avoid touching the sd flow at all. Also add a test to check the correctness.

Concretely, for a batch size of 8 with 128 output tokens, TP=4, for LLama3-70B, the batch latency is
| Without SD | Disable SD | Disable SD after fix |
| -- | -- | -- |
| 5.4 s            |  5.8s            | 5.5s                          |    

Here, without SD means not using sd flag at all. Disable SD means using SD flag, but set `speculative_disable_by_batch_size` smaller than batch size to disable speculative decoding.
After the fix, we are still slower than the original case, this is caused by broadcasting control flow, which will be fixed in future PRs.